Move over exceptions
