+++ edsl/dataset/__init__.py
@@ -3,6 +3,7 @@ from .dataset import Dataset
 from .dataset_operations_mixin import AgentListOperationsMixin
 from .dataset_operations_mixin import ScenarioListOperationsMixin
 from .dataset_operations_mixin import DatasetOperationsMixin
+from .dataset_operations_mixin import ResultsOperationsMixin
 
 __all__ = [
     "Dataset",
