+++ edsl/jobs/__init__.py
@@ -1,4 +1,5 @@
 from .jobs import Jobs
 from .jobs import RunConfig, RunParameters, RunEnvironment
+from .remote_inference import JobsRemoteInferenceHandler
 
 __all__ = ["Jobs"]
